server/api/v1/modules/schemas.py
from pydantic import BaseModel
from typing import List
from ..lessons.schemas import Lesson

class Module(BaseModel):
    title: str
    lessons: List[Lesson] = []

class ModulesResponse(BaseModel):
    modules: List[str]
