server/requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
alembic
asyncpg
